requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.7.1
    # via httpcore
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
django==4.2.4
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
psycopg2-binary==2.9.7
    # via -r requirements.in
python-telegram-bot==20.4
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
urllib3==2.0.4
    # via requests
